fix: resolve some circular dependencies (fully or with `import type`)
