+++ src/components/pages/Triggers/TriggersList/AddTriggerModal/types.ts
@@ -1,0 +1,4 @@
+export enum StepsEnum {
+  condition = 0,
+  action = 1,
+}
