+++ src/models/execution.ts
@@ -1,9 +1,9 @@
-import {Args} from '@models/args';
-import {AssertionResult} from '@models/assertionResult';
-import {EntityMap} from '@models/entityMap';
-import {TestContent} from '@models/test';
-import {TestExecutor} from '@models/testExecutors';
-import {Variables} from '@models/variable';
+import type {Args} from '@models/args';
+import type {AssertionResult} from '@models/assertionResult';
+import type {EntityMap} from '@models/entityMap';
+import type {TestContent} from '@models/test';
+import type {TestExecutor} from '@models/testExecutors';
+import type {Variables} from '@models/variable';
 
 export type ExecutionStatusEnum = 'running' | 'passed' | 'failed' | 'queued' | 'cancelled' | 'aborted';
 export type ExecutionResultOutputTypeEnum = 'text/plain' | 'application/junit+xml' | 'application/json';
