+++ src/models/testSuiteExecution.ts
@@ -1,8 +1,8 @@
-import {EntityMap} from '@models/entityMap';
-import {ExecutionStatusEnum} from '@models/execution';
-import {ObjectRef} from '@models/objectRef';
+import type {EntityMap} from '@models/entityMap';
+import type {ExecutionStatusEnum} from '@models/execution';
+import type {ObjectRef} from '@models/objectRef';
 
-import {TestSuiteStepExecutionResult} from './testSuite';
+import type {TestSuiteStepExecutionResult} from './testSuite';
 
 export type TestSuiteExecution = {
   id: string;
