+++ src/redux/reducers/executorsSlice.ts
@@ -9,7 +9,7 @@ import initialState from '@redux/initialState';
 
 import {isURL} from '@utils/strings';
 
-import {RootState} from '../store';
+import type {RootState} from '../store';
 
 export const executorsSlice = createSlice({
   name: 'executorsSlice',
