+++ src/redux/reducers/testsSlice.ts
@@ -1,10 +1,10 @@
 import {Draft, PayloadAction, createSlice} from '@reduxjs/toolkit';
 
-import {TestWithExecution, TestsState} from '@models/test';
+import type {TestWithExecution, TestsState} from '@models/test';
 
 import initialState from '@redux/initialState';
 
-import {RootState} from '../store';
+import type {RootState} from '../store';
 
 export const testsSlice = createSlice({
   name: 'testsSlice',
